+++ Tests/DataProviderTests.cs
@@ -12,7 +12,7 @@ namespace Tests
         [TestMethod]
         public void TestEmptySystem()
         {
-            StarSystem starSystem = DataProviderService.GetSystemData("Lagoon Sector GW-V b2-6");
+            StarSystem starSystem = DataProviderService.GetSystemData("Lagoon Sector GW-V b2-6", null, null, null);
             Assert.IsNotNull(starSystem.Stations);
         }
     }